requirements.txt
Pillow==2.9.0

# Optional
Werkzeug
django-extensions==1.6.7
#psycopg2==2.5.4
psycopg2==2.6.1
dj-database-url==0.4.0

requests==2.9.1
requests-oauthlib==0.6.0
static3

gunicorn==19.4.5
whitenoise==2.0.6
requests==2.9.1
#oauth2client==2.1.0
oauth2client==1.5.2



celery==3.1.18
djangorestframework==3.3.2
dropbox==3.42
gspread==0.3.0
ipdb==0.8.1
ipython
py-trello==0.4.3
python-nvd3==0.14.2
requests==2.9.1
requests-oauthlib==0.6.0
selenium==2.48.0
xlrd==0.9.4
dj-static
django-compressor==1.4
static3

dj-database-url==0.4.0
Django==1.9.2
gunicorn==19.4.5
psycopg2==2.6.1
whitenoise==2.0.6
requests==2.9.1
#oauth2client==2.1.0
oauth2client==1.5.2
